+++ R/FilePsInput.R
@@ -119,7 +119,7 @@ FilePsInput <- function(id, label = "Upload data (csv/xlsx/sav/sas7bdat/dta)") {
 FilePs <- function(input, output, session, nfactor.limit = 20) {
 
   ## To remove NOTE.
-  BinaryGroupRandom <- variable <- NULL
+  BinaryGroupRandom <- variable <- val_label <- NULL
 
   # The selected file, if any
   userFile <- eventReactive(input$file, {