+++ R/jsSurveyGadget.R
@@ -79,7 +79,7 @@ jsSurveyGadget <- function(data, nfactor.limit = 20) {
                                                      withLoader(DTOutput("untable1"), type="html", loader="loader6"),
                                                      wellPanel(
                                                        h5("Normal continuous variables  are summarized with Mean (SD) and t-test(2 groups) or ANOVA(> 2 groups)"),
-                                                       h5("Non-normal continuous variables are summarized with median [IQR] and kruskal-wallis test"),
+                                                       h5("Non-normal continuous variables are summarized with median [IQR or min,max] and kruskal-wallis test"),
                                                        h5("Categorical variables  are summarized with table")
                                                      )
                                             ),
@@ -87,7 +87,7 @@ jsSurveyGadget <- function(data, nfactor.limit = 20) {
                                                      withLoader(DTOutput("table1"), type="html", loader="loader6"),
                                                      wellPanel(
                                                        h5("Normal continuous variables  are summarized with Mean (SD) and complex survey regression"),
-                                                       h5("Non-normal continuous variables are summarized with median [IQR] and complex sampling rank test"),
+                                                       h5("Non-normal continuous variables are summarized with median [IQR or min,max] and complex sampling rank test"),
                                                        h5("Categorical variables  are summarized with table and svychisq test")
                                                      )
                                             )